requirements.txt
claude-agent-sdk>=0.1.18
flask>=3.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# API Server
fastapi>=0.127.0
uvicorn[standard]>=0.40.0
pydantic>=2.12.5
websockets>=12.0
python-jose[cryptography]>=3.3.0  # JWT token generation and validation
passlib[bcrypt]>=1.7.4  # Password hashing
python-multipart>=0.0.6  # For form data parsing
aiohttp>=3.9.0  # For async HTTP operations and SSE streaming

# PostgreSQL Database
asyncpg>=0.31.0  # High-performance async PostgreSQL driver
psycopg2-binary>=2.9.11  # PostgreSQL adapter (for migrations/sync operations)
sqlalchemy>=2.0.0  # ORM (optional, for future use)

# Sandbox Support
docker>=7.0.0  # Docker SDK for Python (for DockerSandbox)
